Move AddAttributeComponent
